과제 조건들 추가

+++ src/main/java/com/sparta/blogspring/service/BlogService.java
@@ -8,9 +8,7 @@ import com.sparta.blogspring.repository.BlogRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 
 @Service
@@ -22,49 +20,58 @@ public class BlogService {
         this.blogRepository = blogRepository;
     }
 
-    private final Map<Long, Blog> blogList = new HashMap<>();
 
+    //생성하기
     public BlogResponseDto createBlog(BlogRequestDto requestDto) {
         //RequestDto --> Entity
         Blog blog = new Blog(requestDto);
 
         //DB저장
-        Blog savaBlog = blogRepository.save(blog);
-
-        //Entity -> ResponsdeDto
-        BlogResponseDto blogResponseDto = new BlogResponseDto(blog);
-
-        return blogResponseDto;
+        Blog saveBlog = blogRepository.save(blog);
+        //객체 생성하며 반환
+        return new BlogResponseDto(saveBlog);
     }
-
+    //전체조회
     public List<BlogResponseDto> getBlog() {
         //DB조회
         return blogRepository.findAllByOrderByCreatedAtDesc().stream().map(BlogResponseDto::new).toList();
-
     }
 
+    //단건조회
+    public BlogResponseDto getOneBlog(Long id) {
+        //해당 메모의 DB존재 확인
+        Blog blog = findBlog(id);
+        return new BlogResponseDto(blog);
+    }
 
+    //수정하기
     @Transactional
-    public Long updateBlog(Long id, BlogRequestDto requestDto) {
+    public Long updateBlog(Long id, BlogRequestDto requestDto, String password) {
 
         //해당 메모의 DB존재 확인
         Blog blog = findBlog(id);
+        if(password.equals(blog.getPassword())) {
+            blog.update(requestDto);
 
-        //수정
-        blog.update(requestDto);
-
-        return id;
+            return id;
+        } else {
+            throw new IllegalArgumentException("틀린 암호입니다");
+        }
 
     }
-
-    public Long deleteBlog(Long id) {
-
+    //삭제하기
+    public String deleteBlog(Long id, String password) {
 
         Blog blog = findBlog(id);
-        //삭제
-        blogRepository.delete(blog);
 
-        return id;
+        if(password.equals(blog.getPassword())) {
+            blogRepository.delete(blog);
+
+            return "삭제되었습니다.";
+        } else {
+            return "틀린 암호입니다";
+        }
+
     }
 
     //중복되는 매서드 처리 id찾기
